feat: adds ComponentProvider to wrap all providers
Started with an update to dependencies and the required fixes.
This required significant work in
the dev support (tests/stories).
Changes in components now require some global Providers so a new
component provider has been added to wrap them all together.

BREAKING CHANGE: It is now reccommended that you wrap the application with `ComponentsProvider`
replacing the `ThemeProvider`. This includes all required components.

fix #253